pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.1"]
build-backend = "setuptools.build_meta"

[project]
name = "ispdb"
dynamic = ["version"]
description = "ISPDB command line tool"
readme = "README.md"
authors = [
  {name = "Germán Fuentes Capella"}
]
maintainers = [
  {name = "Germán Fuentes Capella"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Email",
    "Topic :: Communications :: Email :: Post-Office :: IMAP",
    "Topic :: Communications :: Email :: Post-Office :: POP3",
]
license = { file = "LICENSE" }

dependencies = [
  "requests>=2.31",
  "dnspython>=2.6.1",
]

[project.optional-dependencies]
dev = [
    "coverage>7.5",  # testing
    "mypy>=1.10",  # type checking
    "pytest>=8.2",  # testing
    "ruff>=0.4",  # linting
    "types-requests>=2.32",
    "requests-mock>=1.12",
    "pytest-mock>=3.14",
]

[project.urls]
bugs = "https://github.com/cloudgerman/ispdb/issues"
homepage = "https://github.com/cloudgerman/ispdb"

[project.scripts]
ispdb = "ispdb.main:main"

[tool.setuptools_scm]
version_file = "src/ispdb/version.py"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.mypy]
files = "."
strict = true
warn_unreachable = true
warn_no_return = true

[tool.ruff]
src = ["src", "tests"]
line-length = 120
